requirements.txt
pytest
pytest-html
playwright
python-dotenv
